backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pydantic
python-jose[cryptography]
bcrypt
python-dotenv
pytest
httpx
email-validator
